backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.4.2
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain-anthropic==0.1.1
langchain-google-genai==0.0.5 